refactor: refactoring nuget-parsers into their own modules
