+++ lib/index.js
@@ -1,6 +1,6 @@
-var path = require('path');
-var dotnetParser = require('./dotnet-parser');
-var paketParser = require('snyk-paket-parser');
+const path = require('path');
+const nugetParser = require('./nuget-parser/index');
+const paketParser = require('snyk-paket-parser');
 
 function determineManifestType(filename) {
   switch (true) {
@@ -25,16 +25,15 @@ function determineManifestType(filename) {
 module.exports = {
   inspect: function (root, targetFile, options) {
     options = options || {};
-    var fileContentPath = path.resolve(root || '.', targetFile || '.');
-    var manifestType;
+    let manifestType;
     try {
       manifestType = determineManifestType(path.basename(targetFile || root));
     } catch (error) {
       return Promise.reject(error);
     }
 
-    var createPackageTree = function (depTree) {
-      var targetFramework = depTree.meta ? depTree.meta.targetFramework : undefined; // TODO implement for paket and more than one framework
+    const createPackageTree = function (depTree) {
+      const targetFramework = depTree.meta ? depTree.meta.targetFramework : undefined; // TODO implement for paket and more than one framework
       delete depTree.meta;
       return {
         package: depTree,
@@ -56,12 +55,10 @@ module.exports = {
       ).then(createPackageTree);
     }
 
-    return dotnetParser.parse(
+    return nugetParser.buildDepTreeFromFiles(
       root,
       targetFile,
       options.packagesFolder,
-      fileContentPath,
       manifestType).then(createPackageTree);
-
   },
 };
