+++ lib/nuget-parser/csproj-parser.js
