+++ lib/nuget-parser/dependency.js
