+++ lib/nuget-parser/dotnet-core-parser.js
@@ -1,7 +1,7 @@
 'use strict';
-var Dependency = require('../dependency');
-var debug = require('debug')('snyk');
-var _ = require('lodash');
+const Dependency = require('../nuget-parser/dependency');
+const debug = require('debug')('snyk');
+const _ = require('lodash');
 
 // TODO: any convention for global vars? (gFreqDeps)
 let freqDeps = {};
@@ -123,38 +123,42 @@ function getTargetObjToRun(manifest) {
   return manifest.targets[selectedTargetKey];
 }
 
+module.exports = {
+  parse: function (tree, fileContent, fileContentParser) {
+    return new Promise(function parseFileContents(resolve) {
+      debug('Trying to parse dot-net-cli manifest');
+      const manifest = fileContentParser.parse(fileContent);
 
-function parse(fileContent, packageTree) {
-  const manifest = JSON.parse(fileContent);
+      if (!manifest.project) {
+        throw new Error('Project field was not found in project.assets.json');
+      }
 
-  if (!manifest.project) {
-    throw new Error('Project field was not found in project.assets.json');
-  }
+      if (manifest.project.version) {
+        tree.version = manifest.project.version;
+      }
 
-  if (manifest.project.version) {
-    packageTree.version = manifest.project.version;
-  }
+      const selectedFrameworkObj = getFrameworkObjToRun(manifest);
+      const selectedTargetObj = getTargetObjToRun(manifest);
 
-  const selectedFrameworkObj = getFrameworkObjToRun(manifest);
-  const selectedTargetObj = getTargetObjToRun(manifest);
+      initFreqDepsDict();
 
-  initFreqDepsDict();
+      const directDependencies = collectFlatList(selectedFrameworkObj.dependencies);
+      debug(`directDependencies: '${directDependencies}'`);
 
-  const directDependencies = collectFlatList(selectedFrameworkObj.dependencies);
-  debug(`directDependencies: '${directDependencies}'`);
+      directDependencies.forEach(function (directDep) {
+        debug(`First order dep: '${directDep}'`);
+        buildTreeRecursive(selectedTargetObj, directDep, tree, 0);
+      });
 
-  directDependencies.forEach(function (directDep) {
-    debug(`First order dep: '${directDep}'`);
-    buildTreeRecursive(selectedTargetObj, directDep, packageTree, 0);
-  });
-
-  if (!_.isEmpty(freqDeps.dependencies.dependencies)) {
-    packageTree.dependencies['freqSystemDependencies'] = freqDeps.dependencies;
-  }
-  // to disconnect the object references inside the tree
-  // JSON parse/stringify is used
-  let pathedTree = JSON.parse(JSON.stringify(packageTree.dependencies));
-  packageTree.dependencies = pathedTree;
-}
-
-module.exports = parse;
+      if (!_.isEmpty(freqDeps.dependencies.dependencies)) {
+        tree.dependencies['freqSystemDependencies'] = freqDeps.dependencies;
+      }
+      // to disconnect the object references inside the tree
+      // JSON parse/stringify is used
+      tree.dependencies = JSON.parse(JSON.stringify(tree.dependencies));
+      resolve(tree);
+    }).catch(function (err) {
+      throw (err);
+    });
+  },
+};
