+++ lib/nuget-parser/nuspec-parser.js
@@ -1,24 +1,24 @@
-var JSZip = require('jszip');
-var fs = require('fs');
-var path = require('path');
-var parseXML = require('xml2js').parseString;
-var Dependency = require('./dependency');
-var _ = require('lodash');
-var debug = require('debug')('snyk');
+const JSZip = require('jszip');
+const fs = require('fs');
+const path = require('path');
+const parseXML = require('xml2js').parseString;
+const Dependency = require('./dependency');
+const _ = require('lodash');
+const debug = require('debug')('snyk');
 
 const targetFrameworkRegex = /([.a-zA-Z]+)([.0-9]+)/;
 
 function parseNuspec(dep, targetFrameworks, sep) {
   return Promise.resolve()
     .then(function () {
-      var pathSep = sep || '.';
-      var nupkgPath =
+      const pathSep = sep || '.';
+      const nupkgPath =
         path.resolve(dep.path, dep.name + pathSep + dep.version + '.nupkg');
-      var nupkgData = fs.readFileSync(nupkgPath);
+      const nupkgData = fs.readFileSync(nupkgPath);
       return JSZip.loadAsync(nupkgData);
     })
     .then(function (nuspecZipData) {
-      var nuspecFile = Object.keys(nuspecZipData.files).find(function (file) {
+      const nuspecFile = Object.keys(nuspecZipData.files).find(function (file) {
         return (path.extname(file) === '.nuspec');
       });
       return nuspecZipData.files[nuspecFile].async('string');
@@ -30,7 +30,7 @@ function parseNuspec(dep, targetFrameworks, sep) {
             return reject(err);
           }
 
-          var ownDeps = [];
+          let ownDeps = [];
           // We are only going to check the first targetFramework we encounter
           // in the future we may want to support multiple, but only once
           // we have dependency version conflict resolution implemented
@@ -109,10 +109,10 @@ function extractDepsForTargetFrameworks(rawDependency, targetFrameworks) {
 }
 
 function extractDepsFromRaw(rawDependencies) {
-  var deps = [];
+  const deps = [];
   _.forEach(rawDependencies, function (dependency) {
     if (dependency && dependency.$) {
-      var dep = new Dependency(dependency.$.id, dependency.$.version);
+      const dep = new Dependency(dependency.$.id, dependency.$.version);
       deps.push(dep);
     }
   });
